+++ index.js
@@ -37,6 +37,6 @@ module.exports = {
   'doublePlaces': require('./validators/doublePlaces'),
   'bigV1Objects': require('./validators/bigV1Objects'),
   'selfIntersectingWaterway': require('./validators/selfIntersectingWaterway'),
-  'separatorToken': require('./validators/separatorToken'),
+  'separatorToken': require('./validators/separatorTokenDestination'),
   'wikiDataNoEnglishName': require('./validators/wikiDataNoEnglishName')
 };
