+++ validators/separatorTokenDestination/index.js
